% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\docType{data}
\name{areas}
\alias{areas}
\title{Motu research areas}
\format{Data frame with columns
\describe{
\item{area}{Area ID}
\item{area_name}{Area name}
\item{area_colour}{Area colour hex code}
}}
\source{
\href{https://motu.nz/resources/working-papers/}{Motu Economic and Public Policy Research}
}
\usage{
data(areas)
}
\description{
Data frame containing research area names and colours.
}
\keyword{datasets}
