% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\docType{data}
\name{authors}
\alias{authors}
\title{Motu working paper authors}
\format{Data frame with columns
\describe{
\item{number}{Working paper number}
\item{author}{Author ID}
}}
\source{
\href{https://motu.nz/resources/working-papers/}{Motu Economic and Public Policy Research}
}
\usage{
data(authors)
}
\description{
Data frame containing paper-author correspondences.
}
\keyword{datasets}
