% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthorship_network.R
\name{coauthorship_network}
\alias{coauthorship_network}
\title{Motu working paper co-authorship network}
\usage{
coauthorship_network(weights = "none")
}
\arguments{
\item{weights}{Character scalar.
Possible values are \code{counts}, \code{newman}, and \code{none} (the default).
See details.}
}
\value{
An igraph graph object.
}
\description{
\code{coauthorship_network} returns the co-authorship network among Motu working paper authors.
}
\details{
\code{coauthorship_network} returns one of three networks, depending on the \code{weights} argument:
\describe{
\item{"counts"}{The network with edge weights equal to the number of working papers co-authored by incident authors.}
\item{"newman"}{The network with edge weights equal to those defined in equation (2) of Newman (2001).}
\item{"none"}{An unweighted network in which authors are adjacent if they have co-authored a working paper.}
}
}
\references{
Newman, M. E. J. (2001).
Scientific collaboration networks: II. Shortest paths, weighted networks, and centrality.
\emph{Physical Review E} 64, 016132.
}
