% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papers.R
\docType{data}
\name{papers}
\alias{papers}
\title{Motu working papers}
\format{Data frame with columns
\describe{
\item{paper}{Working paper number}
\item{title}{Working paper title}
\item{area}{Research area ID}
}}
\source{
\href{https://motu.nz/resources/working-papers/}{Motu Economic and Public Policy Research}
}
\usage{
data(papers)
}
\description{
Data frame containing working paper numbers, titles, and research areas.
}
\keyword{datasets}
